Move callback functions out of signal definitions
